+++ mayan/apps/redactions/layers.py
@@ -2,6 +2,7 @@ from django.utils.translation import ugettext_lazy as _
 
 from mayan.apps.converter.classes import Layer
 
+from .icons import icon_layer_redactions
 from .permissions import (
     permission_redaction_create, permission_redaction_delete,
     permission_redaction_edit, permission_redaction_exclude,
@@ -12,7 +13,7 @@ layer_redactions = Layer(
     empty_results_text=_(
         'Redactions allow removing access to confidential and '
         'sensitive information without having to modify the document.'
-    ), label=_('Redactions'), name='redactions',
+    ), icon=icon_layer_redactions, label=_('Redactions'), name='redactions',
     order=0, permissions={
         'create': permission_redaction_create,
         'delete': permission_redaction_delete,
@@ -20,5 +21,5 @@ layer_redactions = Layer(
         'edit': permission_redaction_edit,
         'select': permission_redaction_create,
         'view': permission_redaction_view
-    }, symbol='highlighter'
+    }
 )